Minor Release Notes: vLMIS version 2.5.1
Date: 29th January, 2016
SONARQUBE: TECHNICAL DEBT RATIO: 4.6%
PHPDOCUMENTOR: IMPLEMENTED

New Feature
--------------
- Add Aggregated Column in All Coverage Reports
- Summary Report in Batch Management for Dry Store Items at all Levels
- Batch Management - Dry Store Products : Implementation of Priority Mechanism / Priority Distribution List
- Enable dry store locations in the receive from supplier form
- User Click Path data insertion function

New Function
---------------
- Received Qty not added in stock in Batch Management and for Issuance

Bug Fixed
---------------
- Disabling Save button on first click on all data entry forms
- Correction of Displayed Formula on National Report
- tOPV-1 and tOPV-0 Consumption auto save fix
- Stock Receive Notification separated from dashboard
- Multiple drafts correction in Issue voucher
- Dropout Formula correction
- Sindh Consumption Form Width expansion
- National Summary and Stock Availability Report adjusted formula as per dashboard
- Stock in Bin - Pagination added
- Cold Chain Capacity Dashboard : removing of zero value bar of non-available vaccine in the cold room
- Reset Button added in Stock Issue & Picking Search Form
- Requirement Issuance Report: Previous year Filter fix
- Stock Issue Search: Cancel Filter correction

+++ vlmis_v2.5.1/code/application/modules/iadmin/controllers/ManageCacheController.php
@@ -1,26 +1,36 @@
 <?php
 
+/**
+ * Iadmin_ManageCacheController
+ *
+ * 
+ *
+ *     Logistics Management Information System for Vaccines
+ * @subpackage Iadmin
+ * @author     Ajmal Hussain <ajmal@deliver-pk.org>
+ * @version    2.5.1
+ */
+
+/**
+ *  Controller for Iadmin Manage Cache
+ */
 class Iadmin_ManageCacheController extends App_Controller_Base {
 
-    public function init() {
-        parent::init();
-    }
-
+    /**
+     * Iadmin_ManageCacheController index
+     */
     public function indexAction() {
-        if ($this->_request->isPost()) {
-            if ($this->_request->getPost()) {
-                $ignoreList = array('.', '..');
-                // path on Test Server
-                // chdir('../../vlmis/');
-                // path for live server
-                chdir('../vlmis/');
-                $arrFiles = scandir('cache', 0);
-                $files = '';
-                foreach ($arrFiles as $file) {
-                    if (!in_array($file, $ignoreList)) {
-                        $files .= $file . '<br />';
-                        unlink("cache/$file");
-                    }
+        if ($this->_request->isPost() && $this->_request->getPost()) {
+            $ignoreList = array('.', '..');
+            // path on Test Server
+            // path for live server
+            chdir('../vlmis/');
+            $arrFiles = scandir('cache', 0);
+            $files = '';
+            foreach ($arrFiles as $file) {
+                if (!in_array($file, $ignoreList)) {
+                    $files .= $file . '<br />';
+                    unlink("cache/$file");
                 }
             }
         }
